models: Move around model definitions in _qbmodel._abstract
The import structure currently makes some things awkward, so break the
abstract `GelModel` apart from the query builder methods interface as
the `AbstractGelModel` and mix the query builder methods into
`BaseGelModel` which is intended to be subclassed by concrete
implementations.  Also, rename `GelLinkModel` and `GelProxyModel` to
`AbstractGelLinkModel` and `AbstractGelProxyModel` for clarity.
